requirements.txt
# pip install --no-cache-dir -r requirements.txt 
albumentations==1.3.0
numpy==1.24.2
opencv-python==4.7.0.68
pandas==1.5.3
Pillow==9.4.0
torch==1.13.1 --extra-index-url https://download.pytorch.org/whl/cpu 
torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.14.1 --extra-index-url https://download.pytorch.org/whl/cpu
tqdm==4.64.1